+++ nextstep-di/src/main/java/nextstep/di/factory/ApplicationContext.java
@@ -2,6 +2,8 @@ package nextstep.di.factory;
 
 import nextstep.annotation.ComponentScan;
 import nextstep.annotation.Configuration;
+import nextstep.di.factory.exception.DuplicateBeanException;
+import nextstep.di.factory.exception.NotFoundComponentScanException;
 import nextstep.stereotype.Controller;
 import nextstep.stereotype.Repository;
 import nextstep.stereotype.Service;
@@ -32,21 +34,6 @@ public class ApplicationContext {
             log.error("compinent Scan을 찾을 수 없습니다.");
             throw new NotFoundComponentScanException();
         }
-    }
-
-    /**
-     * bean 클래스가 중복되었을 때의 처리에 대해 고민해 보면 어떨까요?
-     *
-     * 중복된 처리에 대한 테스트도 추가되면 좋아 보입니다.
-     */
-    public Map<Class<?>, BeanDefinition> scanBeans2() {
-        Map<Class<?>, BeanDefinition> preInitiateBeans = new HashMap<>();
-
-        for (BeanScanner beanScanner : beanScanners) {
-            preInitiateBeans.putAll(beanScanner.scanBeans());
-        }
-
-        return preInitiateBeans;
     }
 
     public Map<Class<?>, BeanDefinition> scanBeans() {
