+++ nextstep-di/src/main/java/nextstep/di/factory/exception/InvalidBeanException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class InvalidBeanException extends RuntimeException {
     public InvalidBeanException() {
