+++ nextstep-di/src/main/java/nextstep/di/factory/exception/NotFoundConstructorException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class NotFoundConstructorException extends RuntimeException {
     public NotFoundConstructorException() {
